signal_dispatcher/requirements.txt

google-cloud-bigquery>=3.0.0
pydantic>=2.0.0
requests>=2.28.0
facebook-business>=18.0.0
